Gather new format data

+++ SensorValue.cs
@@ -8,14 +8,13 @@ namespace Medic
 {
     class SensorValue
     {
-        public float x { get; set; }
-        public float y { get; set; }
-        public float z { get; set; }
+        public short x { get; set; }
+        public short y { get; set; }
+        public short z { get; set; }
 
         public String toString()
         {
             return "X:" + x + " Y:" + y + " Z:" + z;
         }
-
     }
 }
